Split prog in 3
